Moving tachyon.client.worker classes into tachyon.client.block.
