+++ clients/unshaded/src/main/java/tachyon/client/block/BlockStoreContext.java
@@ -26,7 +26,6 @@ import com.google.common.base.Throwables;
 
 import tachyon.client.ClientContext;
 import tachyon.client.ClientUtils;
-import tachyon.client.worker.BlockWorkerClient;
 import tachyon.exception.ExceptionMessage;
 import tachyon.exception.PreconditionMessage;
 import tachyon.thrift.WorkerInfo;
