+++ clients/unshaded/src/main/java/tachyon/client/block/BlockWorkerClientPool.java
@@ -23,7 +23,6 @@ import org.slf4j.LoggerFactory;
 import tachyon.Constants;
 import tachyon.client.ClientContext;
 import tachyon.client.ClientUtils;
-import tachyon.client.worker.BlockWorkerClient;
 import tachyon.resource.ResourcePool;
 import tachyon.worker.NetAddress;
