+++ clients/unshaded/src/main/java/tachyon/client/block/LocalBlockInStream.java
@@ -23,7 +23,6 @@ import javax.annotation.concurrent.NotThreadSafe;
 import com.google.common.io.Closer;
 
 import tachyon.client.ClientContext;
-import tachyon.client.worker.BlockWorkerClient;
 import tachyon.exception.ExceptionMessage;
 import tachyon.exception.TachyonException;
 import tachyon.thrift.LockBlockResult;
