+++ clients/unshaded/src/main/java/tachyon/client/block/LocalBlockOutStream.java
@@ -27,7 +27,6 @@ import com.google.common.io.Closer;
 
 import tachyon.Constants;
 import tachyon.client.ClientContext;
-import tachyon.client.worker.BlockWorkerClient;
 import tachyon.exception.ExceptionMessage;
 import tachyon.exception.TachyonException;
 import tachyon.util.io.FileUtils;
