+++ clients/unshaded/src/main/java/tachyon/client/block/WorkerClientHeartbeatExecutor.java
@@ -13,7 +13,9 @@
  * the License.
  */
 
-package tachyon.client.worker;
+package tachyon.client.block;
+
+import javax.annotation.concurrent.ThreadSafe;
 
 import com.google.common.base.Preconditions;
 
@@ -23,6 +25,7 @@ import tachyon.heartbeat.HeartbeatExecutor;
  * Session client sends periodical heartbeats to the worker it is talking to. If it fails to do so,
  * the worker may withdraw the space granted to the particular session.
  */
+@ThreadSafe
 final class WorkerClientHeartbeatExecutor implements HeartbeatExecutor {
   private final BlockWorkerClient mBlockWorkerClient;
