+++ integration-tests/src/test/java/tachyon/worker/DataServerIntegrationTest.java
@@ -42,7 +42,7 @@ import tachyon.client.block.BlockMasterClient;
 import tachyon.client.block.BlockStoreContext;
 import tachyon.client.file.FileSystem;
 import tachyon.client.file.URIStatus;
-import tachyon.client.worker.BlockWorkerClient;
+import tachyon.client.block.BlockWorkerClient;
 import tachyon.conf.TachyonConf;
 import tachyon.exception.ConnectionFailedException;
 import tachyon.exception.TachyonException;
